Z-lift before servo retraction
Added Z-Lift before servo retraction for G28 and G29.
Also: For G29 Auto bed leveling, it no longer retracts the servo before moving to the next probing position, instead it retracts at the end of the sequence.

This is obviously a crude implementation. I am not a skilled programmer, but it works on my machine.